bkp_requirements.txt
aenum==3.1.16
ai-api-client-sdk==2.6.1
ai-core-sdk==2.6.2
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.37.0
botocore==1.37.1
cachetools==5.5.2
camelot-py==1.0.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.2
dacite==1.9.2
dataclasses-json==0.6.7
DBUtils==3.1.0
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Limiter==3.12
frozenlist==1.6.0
fuzzywuzzy==0.18.0
generative-ai-hub-sdk==4.12.4
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-aiplatform==1.83.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hdbcli==2.24.24
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jpype1==1.5.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-aws==0.2.15
langchain-community==0.3.24
langchain-core==0.3.60
langchain-google-vertexai==2.0.15
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
limits==5.2.0
logger==1.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.79.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optional==0.0.1
ordered-set==4.1.0
orjson==3.10.18
overloading==0.5.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
pyhumps==3.8.0
pypdf==5.5.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.11.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0